Add spec test for validate_bool function
This is an interesting spec test for module developers.
It illustrates how to cause Puppet to test the function
from the Puppet DSL rather than the Ruby DSL, fully
exercising the system from the perspective of the end
user.

(Note how Puppet[:code] is set, then the scope reset, then
the compile method called.)

Paired-with: Dan Bode <dan@puppetlabs.com>