(#12357) Fix root_home fact on Windows
Without this patch the root_home fact fails on windows.  This patch
fixes the problem by implementing a code block confined to windows for
the fact's value resolution.

The implementation is as follows, looking entirely to the environment:

 * Use %ProgramData% if it is set, otherwise...
 * Use %LOCALAPPDATA% if it is set, otherwise...
 * Use %TEMP% if it is set, otherwise...
 * nil

There is no concept of a "home directory for the root user" on windows,
so we're defaulting to the closest thing we have available, the
%ProgramData% folder on modern windows operating systems.