Add has_interface_with() loopback specification for MacOSX and Linux
Without this patch the specification of how the has_interface_with()
function is not working and does not appear to have a good example of
how to proceed.  This is a problem because the function need to work
well on all platforms that Puppet and stdlib support.

This patch helps to fix the problem by providing an example of how to
mock lookupvar in a way that provides fact values for different
operating systems.

This also allows us to specify the behavior for Mac OS X and validate
the specification on a Linux host, and vice-versa.

The tests should look like:

rspec --format d spec/unit/puppet/parser/functions/has_interface_with_spec.rb

    function_has_interface_with
      On Mac OS X Systems
        should have loopback (lo0)
        should not have loopback (lo)
      On Linux Systems
        should have loopback (lo)
        should not have loopback (lo0)

    Finished in 0.04262 seconds
    4 examples, 0 failures