Changes to return false if no _interfaces_ fact
Instead of raising a parse error if there is no _interfaces_ fact, we just
return false--if there aren't any interfaces, then there cannot be any
attributes of interfaces.

This is perhaps not as obvious of a choice as it may seem, as the only
cases I can think of where there is no _interfaces_ fact are a system in
error (or an incomplete testing environment). In those cases, it might
be better to puke than to just silently fail. But it's not inconceivable
that a system might exist where no _interfaces_ fact is defined, so I am
erroring on the side of obvious and logical instead of helpful.