Disable tests that fail on 2.6.x due to #15912
In Puppet 2.6.x there is a bug where a function may be incorrectly detected as
an rvalue when it is not, or not detected when it is. This means that in tests
the correct syntax for calling a function will be rejected. This disables
those tests on 2.6.x, as there is no straightforward way to write them to be
compatible with both 2.6.x and newer versions of Puppet.

Conflicts:
	spec/unit/puppet/parser/functions/getvar_spec.rb
	spec/unit/puppet/parser/functions/has_key_spec.rb
	spec/unit/puppet/parser/functions/merge_spec.rb

This patch was constructed by cherry-picking e27eccb and resolving the
merge conflicts to only include the `pending` statements.  This resolves
the problem by disabling these tests in Puppet 2.6.