Add pry() function
It's often useful to drop into a debugger shell in the middle of a catalog compilation to see exactly which resources have been added, what class parameters are set to, and which variables have actually been evaluated as expected.

Traditionally we resort to `$foo = inline_template("<%= require 'pry' ; binding.pry %>")` but in the interest of puppet code golfing would like to propose `pry()` as a shorter alternative.