New function: hash_fetch, multilevel fetch on a nested hash
This function is similar to a fetch function on a Hash object with the
exception that it works on a nested hash.

It tries to retrieve the value to the specified key. If the key does
not exist it will return the specified default value. If key is an
array of keys then this array is used to walk through a nested hash.

If it any time a key does not exist we return the default value.
If no default value is specified then undef is used as default.

Change-Id: If33fe5870e67ac6994211f29fe3e7a08c1cfa24f