(doc) Update to point to Jira
Since we've moved from Redmine to Jira the links need to be updated so
that people know where to look for issues.

At the moment stdlib is being tracked with puppet in the PUP project.
This doesn't seem like a good, long term solution, but it is where we
are right now.