Add string_rand() function
string_rand() behaves mostly like fqdn_rand(). The difference being the
fqdn is not used as a seed to generate a random number, rather a string
(or interger) arguement is used.

This function can be useful in providing a random number that can be
used as an identifier for clustered services.