Updated file_line docs for accuracy and non-badness.
The `file_line` with `match` example was broken, because the `match`
didn't match the `line`.

In addition, it was a bad example, because it shows using `file_line` to
modify `/etc/fstab`, which should really be done with the motunt
provider.

In addition, the text didn't explain fully how to use match.