(#MODULES-865) validate_re cannot match numeric regex in all cases
The validate_re function now tries to convert it's first argument to
a string. This will ensure that numeric, boolean, symbols, arrays and hashes can be matched
by a regex since regex matching can only happen on strings.

validate_re(1, '^\d+$') would previously fail because the first argument is a Fixnum.
This patch ensures the first argument is cast to a string so this would match as one would expect.