(MODULES-2031) Add 'url_encode' function to stdlib
We recently upgraded our stdlib module to 4.6 and discovered that one of
our modules depended on legacy behavior of the uriescape function that
had changed significantly along the way:

https://github.com/puppetlabs/puppetlabs-stdlib/pull/164
(commit: b2e23dc65bd9851dcb6ad60ffca1acbc70b617e1)

Before PR#164, the effective behavior of uriescape was to replace
characters that are illegal or significant within URIs with
percent-delimited, hex-encoded equivalents, as is necessary when
encoding a free-text UTF-8 string within a URL path segment ( / "text"
    / ) such that any forward-slashes, ampersands, question marks, and
octothorpes embedded within the text do not change the structure of
the URL from the perspective of a URL parser.

After PR#164, the effective behavior of uriescape changed, for the
most part, to minimally escape a complete URL (agnostic of its parts)
in such a way to make it safe to pass as a double-quoted argument to a
shell command like curl, to prevent command injection.

We still need the legacy behavior for building URIs, such as JDBC URLs
in config files that embed a password value from hiera containing
URI-significant characters like '&' or '?'.

Since the uriescape function has been adjusted to match the behavior
of the Ruby URI.escape function, a simple enhancement would be to
expose one of the recommended alternative Ruby functions
(ERB::Util.url_encode), as a distinct puppet function.