Add ensure_latest function.
This provides a convenient way to idempotently ask for a package, but also
makes it effortless to always get the latest version.  This prevents the
"version sprawl" enabled by ensure_packages, where machines in a fleet that
are provisioned over time get different package versions.  At the same time,
we are maintaining the familiar interface provided by ensure_packages.