Add function destringify
Loads a fact or other variable, and in case of a stringified fact
containing an
array, string, or hash, it returns the data in the corresponding
native data type.

Puppet 3.x will sometimes convert all fact values to strings
(e.g. "false" instead of false), depending on the stringify_facts
setting
and the installed Facter version.

If you’re writing code that might be used with pre-4.0 versions of
Puppet,
you’ll need to take extra care when dealing with structured facts.

For example:

if the '::processors' fact is stringified as follow:
  "{\"count\"=>8, \"speed\"=>\"2.7 GHz\"}"

  $my_destringified_processors_fact = destringify(::processors)
    => {"count"=>8, "speed"=>"2.7 GHz"}

This could also be done with following puppet/stdlib functions:

  $_fact = $::my_structured_fact
  if is_string($_fact) {
    $_yaml = regsubst($_fact,'=>',': ', 'G')
    $_destringified_fact = parseyaml($_yaml)
  } else {
    $_destringified_fact = $_fact
  }

But using this function it will be reduced to:

$_destringified_fact = destringify($::my_structured_fact)