Empower sort to (recursively) sort a hash
This only allows for sorting based on the key and will throw an error if
the keys are of different types. It's mostly meant to help out with
cases where we're dealing with JSON as input or output and want to
guarantee that the order of things don't change between Puppet runs,
potentially breaking idempotency.