(MODULES-3829) Make ensure_packages work with < 2.0
Prior to this commit, if a hash was passed in as an
argument to the ensure_packages function a method of
hash duplication would be used that is not supported
with versions of ruby older than 2.0. In order to
ensure the method is compatible with older ruby versions,
switch to a different method of duplication.

Additionally add tests to prevent further regressions.