Revert "Allow use of fact() on other hashes"
This reverts commit 409a974095a3f5b637e091494b5d14b451c5de78.

After thinking about this, use case, and function naming, I think it's
probably best for now to keep things simple and let `fact()` be
single-purpose for looking up facts with no potentially confusing
extensions. The only use case where the name makes sense is where it's
being used on the `$facts` hash, and I think we'd rather in that
circumstance promote the raw use of `fact()`.