(MODULES-5651) append_on_no_match method fix
This renames append_on_no_match to simply append to relieve lots of double negatives and adds a fix to the method. Previously, if the param was set to false then it always returns true before the replace param is evaluated, so nothing happens. This being the case, I added an additional condition , so that it only acts when there's no match.