(MODULES-5651) Do not append infinitely
https://tickets.puppetlabs.com/browse/MODULES-5003 gave rise to
https://github.com/puppetlabs/puppetlabs-stdlib/pull/788 and
https://github.com/puppetlabs/puppetlabs-stdlib/pull/794 which caused
different behavior based on whether the line value was matched by the
match regex or not.

The change in behavior was both breaking and broken, though that was
hard to tell because the behavior was ill-described in general.

[bugfix] This commit resolves the breaking behavior by reverting the
behavior of "replacing matches when a line matching `line` exists even
when `multiple` is set to `true`".

[feature] This commit adds a new parameter to make file_line replace all
matches universally with the `line` value, even when the line exists
elsewhere in the file. This feature only affects modifying multiple
lines in a file when the `line` value already exists.

[bugfix] This commit more strictly defines the various interactions of
`ensure`, `match`, `append_on_no_match`, `replace`, `multiple`, and
`replace_all_matches_not_matching_line`. It also more clearly documents
and tests these interactions.