(MAINT) Test Travis bundler fail theory
As discussed in MODULES-6339, there are some ongoing problems
with Travis, bundler, and ruby versions. This commit is
an experiment to see what versions of bundler come with
which images and to determine whether or not we can remove
the update line or if we should pin to a particular version.