Make any2array return empty array on empty string
any2array currently returns an array containing the empty string, when
given the empty string.

Example:
```
any2array('').each |$e| {
  notice('I\'m not empty!')
}
```

This is counter intuitive, as Puppet considers the empty string as
Strings nil value, for which the equivalent Array is the empty array.

This especially apparent when doing a lookup in a map with a non
existing key.

Example:
```
$config = {}
any2array($config['test']).each |$e| {
  notice('I\'m not empty!')
}
```

Both these cases print the string "I'm not empty", this PR fixes that.